refactor: createManualSystem 분리
